requirements.txt
asgiref==3.7.2
click==7.1.2
click-didyoumean==0.3.0
Django==4.2.7
djangorestframework==3.14.0
p==1.4.0
pytz==2023.3.post1
sqlparse==0.4.4
